requirements/common.txt
Django==3.2

boto3==1.24.13
dj-database-url==0.5.0
djoser==2.1.0
django-anymail[amazon_ses]==8.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.1.5
django-filter==22.1
django-model-utils==4.2.0
django-safedelete==1.2.2
django-simple-history==3.1.1
django-storages==1.12.3
django-storage-url==0.8.1
django-tenant-schemas==1.11.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
faker==13.13.0
ipython==8.4.0
jsonschema==4.6.0
markdown==3.3.7
model-bakery==1.5.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pyyaml==6.0
pytest-django==4.5.2
pytest-cov==3.0.0
pytest-mock==3.7.0
pytest-xdist==2.5.0
sentry-sdk==1.5.12
uritemplate==4.1.1
whitenoise==6.2.0
